get datalink functions and resolve conflicts

* master: (33 commits)
  [supervision] color regex on "error:" and "error "
  [telemetry] rename device.h link_device.h to match structure name
  fix_code_style.sh: align pointer with name
  [telemetry] rename 'device' to 'link_device'
  [build] don't print infos on stderr
  fix travis config
  [airborne] get rid of some warnings
  [test] update compile_all_aircrafts test
  [conf.xml] update to new settings_modules
  [ext][stm32] update libopencm3
  [GCS] reword maps fill mouse over
  [replay] filter messages to remove warnings during replay
  more bug fix
  missing parenthesis
  [gen_fligtplan] 'call' statement can be configured to loop and/or break
  fix some bug reported by TravisCI
  last minute fix
  [telemetry] fixing and cleaning a few stuff
  [telemetry] add support for ivy_transport
  [rotorcraft] get rid of some warnings in ardrone2_sdk
  ...

+++ sw/airborne/arch/stm32/peripherals/ms2100_arch.c
@@ -69,7 +69,7 @@ void ms2100_reset_cb( struct spi_transaction * t __attribute__ ((unused)) ) {
   int32_t end_cpu_ticks = systick_get_value() - dt_ticks;
   if (end_cpu_ticks < 0)
     end_cpu_ticks += systick_get_reload();
-  while (systick_get_value() > end_cpu_ticks)
+  while (systick_get_value() > (uint32_t)end_cpu_ticks)
     ;
 
   Ms2100Reset();
